Refactor: simplificando Reserva com construtor primário e nullable types

+++ Models/Suite.cs
@@ -2,16 +2,14 @@ namespace Sistema_de_Hospedagem.Models
 {
     public class Suite
     {       
-        public Suite(string tipoSuite, int capacidade, int valorDiaria) 
+        public Suite(string tipoSuite, int capacidade, decimal valorDiaria) 
         {
             TipoSuite = tipoSuite;
             Capacidade = capacidade;
             ValorDiaria = valorDiaria;
         }
-
         public string? TipoSuite { get; set; }
         public int Capacidade { get; set; }
         public decimal ValorDiaria { get; set; }
-
     }
 }
